Fix CORS issues in Azure dev environment, Bicep templates and add instructions for setting up DB roles
